moved a function common to both packages acquired a lock before writing to file
