+++ aitelemetry/api.go
@@ -1,6 +1,7 @@
 package aitelemetry
 
 import (
+	"github.com/Azure/azure-container-networking/common"
 	"github.com/microsoft/ApplicationInsights-Go/appinsights"
 )
 
@@ -23,32 +24,8 @@ type TelemetryHandle struct {
 	telemetryConfig       *appinsights.TelemetryConfiguration
 	AppName               string
 	AppVersion            string
-	metadata              Metadata
+	metadata              common.Metadata
 	diagListener          appinsights.DiagnosticsMessageListener
 	client                appinsights.TelemetryClient
 	isMetadataInitialized bool
-}
-
-// Metadata retrieved from wireserver
-type Metadata struct {
-	Location             string `json:"location"`
-	VMName               string `json:"name"`
-	Offer                string `json:"offer"`
-	OsType               string `json:"osType"`
-	PlacementGroupID     string `json:"placementGroupId"`
-	PlatformFaultDomain  string `json:"platformFaultDomain"`
-	PlatformUpdateDomain string `json:"platformUpdateDomain"`
-	Publisher            string `json:"publisher"`
-	ResourceGroupName    string `json:"resourceGroupName"`
-	Sku                  string `json:"sku"`
-	SubscriptionID       string `json:"subscriptionId"`
-	Tags                 string `json:"tags"`
-	OSVersion            string `json:"version"`
-	VMID                 string `json:"vmId"`
-	VMSize               string `json:"vmSize"`
-	KernelVersion        string
-}
-
-type metadataWrapper struct {
-	Metadata Metadata `json:"compute"`
 }
