--- aitelemetry/utils.go
@@ -1,83 +1,0 @@
-package aitelemetry
-
-import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"net"
-	"net/http"
-	"time"
-
-	"github.com/Azure/azure-container-networking/log"
-)
-
-const (
-	metadataURL           = "http://169.254.169.254/metadata/instance?api-version=2017-08-01&format=json"
-	httpConnectionTimeout = 10
-	headerTimeout         = 20
-)
-
-// getHostMetadata - retrieve metadata from host
-func getHostMetadata() (Metadata, error) {
-	content, err := ioutil.ReadFile(metadataFile)
-	if err == nil {
-		var metadata Metadata
-		if err = json.Unmarshal(content, &metadata); err == nil {
-			return metadata, nil
-		}
-	}
-
-	log.Printf("[Telemetry] Request metadata from wireserver")
-
-	req, err := http.NewRequest("GET", metadataURL, nil)
-	if err != nil {
-		return Metadata{}, err
-	}
-
-	req.Header.Set("Metadata", "True")
-
-	client := &http.Client{
-		Transport: &http.Transport{
-			DialContext: (&net.Dialer{
-				Timeout: time.Duration(httpConnectionTimeout) * time.Second,
-			}).DialContext,
-			ResponseHeaderTimeout: time.Duration(headerTimeout) * time.Second,
-		},
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return Metadata{}, err
-	}
-
-	defer resp.Body.Close()
-
-	metareport := metadataWrapper{}
-
-	if resp.StatusCode != http.StatusOK {
-		err = fmt.Errorf("[Telemetry] Request failed with HTTP error %d", resp.StatusCode)
-	} else if resp.Body != nil {
-		err = json.NewDecoder(resp.Body).Decode(&metareport)
-		if err != nil {
-			err = fmt.Errorf("[Telemetry] Unable to decode response body due to error: %s", err.Error())
-		}
-	} else {
-		err = fmt.Errorf("[Telemetry] Response body is empty")
-	}
-
-	return metareport.Metadata, err
-}
-
-// saveHostMetadata - save metadata got from wireserver to json file
-func saveHostMetadata(metadata Metadata) error {
-	dataBytes, err := json.Marshal(metadata)
-	if err != nil {
-		return fmt.Errorf("[Telemetry] marshal data failed with err %+v", err)
-	}
-
-	if err = ioutil.WriteFile(metadataFile, dataBytes, 0644); err != nil {
-		log.Printf("[Telemetry] Writing metadata to file failed: %v", err)
-	}
-
-	return err
-}
